some minor fixes
